general overhaul
Alright, let's go over this.

I've created some classes to help with Unity autosplitters at [just-ero/AutoSplitHelp](https://github.com/just-ero/AutoSplitHelp). To test this autosplitter locally, you must download `mono.cs` and `mono_helpers.cs` and put it into the `Components` folder wherever you saved LiveSplit. For further distribution to other users, simply change the `LiveSplit.AutoSplitters.xml` file to look like [this](https://github.com/just-ero/AutoSplitHelp/blob/main/mono/LiveSplit.AutoSplitters-Example_PR.xml).

## What does this do?
The main idea is to stop using pointers. Unity uses Mono, which loads all classes in static memory. This means all classes can be founds via their name (or token), along with all of their fields (and their offsets). A class' static fields can then be accessed via a static address.  
`mono.cs` and `mono_helpers.cs` do this for the user, who then only has to specify which classes to get.

## Specifically what was changed?
`start {}`: Since the game handles the in-game time a bit weirdly (arbitrarily starts the timer before the first loading screen and then waits for it to finish), we first wait for the `LevelManagerScript.loadQueue` to empty, setting a flag to allow the starting of the timer when the in-game time finally begins counting up again.

`split {}`: Split either when the player's rank increases or they collect the final tape in the final level. ***This must be tested!*** I am not good, nor dedicated enough to find out if this works the way it should. This value might count up throughout the entire playthrough, meaning it does not reset when a new level is entered. If this is the case, simply store the tape count when entered level 5 and check that `value + 3` against `current.tapesCollected`.

`reset {}`: The game stores the `DateTime.Now` value of when the last run was started. When this value increases, a new run has been started.